fix error when attempting to delete tdmp folders when they dont exist

+++ src/TeardownMultiplayerLauncher/Core/Services/TeardownMultiplayerUpdateService.cs
@@ -10,13 +10,17 @@ namespace TeardownMultiplayerLauncher.Core.Services
 {
     internal class TeardownMultiplayerUpdateService
     {
+        private readonly LauncherState _state;
         private readonly IPackageResolver _packageResolver;
-        private readonly TeardownMultiplayerUpdateState _state;
 
-        public TeardownMultiplayerUpdateService(TeardownMultiplayerUpdateState state)
+        public TeardownMultiplayerUpdateService(LauncherState state)
         {
-            _packageResolver = new GithubPackageResolver(state.GitHubRepositoryOwner, state.GitHubRepositoryName, state.GitHubAssetFileNamePattern);
             _state = state;
+            _packageResolver = new GithubPackageResolver(
+                state.TeardownMultiplayerUpdateState.GitHubRepositoryOwner,
+                state.TeardownMultiplayerUpdateState.GitHubRepositoryName,
+                state.TeardownMultiplayerUpdateState.GitHubAssetFileNamePattern
+            );
         }
 
         public async Task SetUpLatestReleaseAsync(string teardownDirectory)
@@ -32,10 +36,10 @@ namespace TeardownMultiplayerLauncher.Core.Services
                 var zipFilePath = await DownloadReleaseZipAsync(latestReleaseVersion);
                 await UninstallCurrentReleaseAsync(teardownDirectory);
                 await InstallLatestReleaseFromZipAsync(zipFilePath, teardownDirectory);
-                _state.InstalledVersion = latestReleaseVersion.ToString();
+                _state.TeardownMultiplayerUpdateState.InstalledVersion = latestReleaseVersion.ToString();
             }
 
-            _state.LastCheckDateTimeUtc = DateTime.UtcNow;
+            _state.TeardownMultiplayerUpdateState.LastCheckDateTimeUtc = DateTime.UtcNow;
         }
 
         /// <summary>
@@ -43,12 +47,12 @@ namespace TeardownMultiplayerLauncher.Core.Services
         /// </summary>
         private bool IsOnCooldown()
         {
-            return _state.LastCheckDateTimeUtc.HasValue && (DateTime.UtcNow - _state.LastCheckDateTimeUtc) < _state.CheckCooldownDuration;
+            return _state.TeardownMultiplayerUpdateState.LastCheckDateTimeUtc.HasValue && (DateTime.UtcNow - _state.TeardownMultiplayerUpdateState.LastCheckDateTimeUtc) < _state.TeardownMultiplayerUpdateState.CheckCooldownDuration;
         }
 
         private bool IsLatestReleaseNewerThanInstalledVersion(Version latestReleaseVersion)
         {
-            return !string.Equals(latestReleaseVersion.ToString(), _state.InstalledVersion, StringComparison.Ordinal);
+            return !string.Equals(latestReleaseVersion.ToString(), _state.TeardownMultiplayerUpdateState.InstalledVersion, StringComparison.Ordinal);
         }
 
         private async Task<string> DownloadReleaseZipAsync(Version releaseVersion)
@@ -63,17 +67,25 @@ namespace TeardownMultiplayerLauncher.Core.Services
             await Task.Run(() =>
             {
                 // Remove TDMP folders. This is required in case if files was removed or something unexpected happened.
-                Directory.Delete(Path.Combine(teardownDirectory, "mods/TDMP"), true);
-                Directory.Delete(Path.Combine(teardownDirectory, "data/TDMP"), true);
-
-                foreach (var filePath in _state.InstalledFilePaths)
+                var tdmpModsDirectory = Path.Combine(teardownDirectory, "mods/TDMP");
+                var tdmpDataDirectory = Path.Combine(teardownDirectory, "data/TDMP");
+                if (Directory.Exists(tdmpModsDirectory))
                 {
-                    if (File.Exists(filePath))
+                    Directory.Delete(tdmpModsDirectory, true);
+                }
+                if (Directory.Exists(tdmpDataDirectory))
+                {
+                    Directory.Delete(tdmpDataDirectory, true);
+                }
+
+                foreach (var filePath in _state.TeardownMultiplayerUpdateState.InstalledFilePaths)
+                {
+                    if (File.Exists(filePath) && filePath.StartsWith(teardownDirectory))
                     {
                         File.Delete(filePath);
                     }
                 }
-                _state.InstalledFilePaths.Clear();
+                _state.TeardownMultiplayerUpdateState.InstalledFilePaths.Clear();
             }).ContinueWith(task =>
             {
                 if (task.IsFaulted)
@@ -93,7 +105,7 @@ namespace TeardownMultiplayerLauncher.Core.Services
                     var isFile = !string.IsNullOrWhiteSpace(zipEntry.Name); // Folders don't have "Name" populated, only files do.
                     if (isFile)
                     {
-                        _state.InstalledFilePaths.Add(Path.Combine(teardownDirectoryPath, zipEntry.FullName));
+                        _state.TeardownMultiplayerUpdateState.InstalledFilePaths.Add(Path.Combine(teardownDirectoryPath, zipEntry.FullName));
                     }
                 }
                 ZipFile.ExtractToDirectory(zipFilePath, teardownDirectoryPath, true);
